Sv/modify sp scripts

See merge request cpsc319-2018w2/jabc/oecm/jabc-hr!20

+++ db/sql-scripts/Create_SP_Document.sql
@@ -1,6 +1,11 @@
 -- STORED PROCEDURES FOR DocumentService
 
--- Create new support doc type
+
+-- -----------------------------------------------------
+-- procedure create_doc_type
+--    - create a doc type, only if its type name isn't
+--    - in use
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS create_doc_type;
 
 DELIMITER //
@@ -9,59 +14,117 @@ CREATE PROCEDURE `create_doc_type` (IN type_name VARCHAR(48)
 , IN path VARCHAR(512)
 )
 BEGIN
-    INSERT INTO `TYPE` (TYPE_NAME, PATH)
-    VALUES (type_name, path);
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(TYPE_NAME) INTO checker FROM `TYPE` WHERE `TYPE`.TYPE_NAME = type_name;
+
+    IF checker = 1 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Document type name is already in use.';
+    ELSE
+      INSERT INTO `TYPE` (TYPE_NAME, PATH)
+      VALUES (type_name, path);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- delete support doc
+-- -----------------------------------------------------
+-- procedure delete support doc
+--    - delete a support doc, only if it exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS delete_support_doc;
 
 DELIMITER //
 
 CREATE PROCEDURE `delete_support_doc` (IN support_doc_id INT)
-
 BEGIN
-    DELETE FROM `SUPPORT_DOC`
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(SUPPORT_DOC_ID) INTO checker
+    FROM `SUPPORT_DOC`
     WHERE `SUPPORT_DOC`.SUPPORT_DOC_ID = support_doc_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Support document does not exist.';
+    ELSE
+      DELETE FROM `SUPPORT_DOC`
+      WHERE `SUPPORT_DOC`.SUPPORT_DOC_ID = support_doc_id;
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- get_support_docs(IN id INT)
+-- -----------------------------------------------------
+-- procedure get_support_docs
+--    - get support docs for an employee, if that
+--    - employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_support_docs;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_support_docs` (IN id INT)
+CREATE PROCEDURE `get_support_docs` (IN employee_id INT)
 BEGIN
-    SELECT *
-    FROM SUPPORT_DOC
-    WHERE EMPLOYEE_ID = id;
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(EMPLOYEE_ID) INTO checker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+    ELSE
+      SELECT *
+      FROM SUPPORT_DOC
+      WHERE EMPLOYEE_ID = employee_id;
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Get support doc type
+-- -----------------------------------------------------
+-- procedure get_doc_type
+--    - get the doc type given its id, if it exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_doc_type;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_doc_type` (IN id INT)
+CREATE PROCEDURE `get_doc_type` (IN type_id INT)
 BEGIN
-    SELECT *
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(TYPE_ID) INTO checker
     FROM `TYPE`
-    WHERE TYPE_ID = id;
+    WHERE `TYPE`.TYPE_ID = type_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Document type does not exist.';
+    ELSE
+      SELECT *
+      FROM `TYPE`
+      WHERE TYPE_ID = id;
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Get support doc types
+-- -----------------------------------------------------
+-- procedure get_all_doc_types
+--    - get all the document types
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_all_doc_types;
 
 DELIMITER //
@@ -75,7 +138,10 @@ END //
 DELIMITER ;
 
 
--- Update given support doc
+-- -----------------------------------------------------
+-- procedure update_support_doc
+--    - update the given support doc, if it exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS update_support_doc;
 
 DELIMITER //
@@ -90,16 +156,31 @@ CREATE PROCEDURE `update_support_doc` (IN support_doc_id INT
 , IN description VARCHAR(512)
 )
 BEGIN
-    UPDATE SUPPORT_DOC
-    SET EMPLOYEE_ID = employee_id, TYPE_ID = type_id, CREATED_DATE = created_date, DUE_DATE = due_date,
-    EXPIRY_DATE = expiry_date, PATH = path, DESCRIPTION = description
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(SUPPORT_DOC_ID) INTO checker
+    FROM `SUPPORT_DOC`
     WHERE `SUPPORT_DOC`.SUPPORT_DOC_ID = support_doc_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Support document does not exist.';
+    ELSE
+      UPDATE SUPPORT_DOC
+      SET EMPLOYEE_ID = employee_id, TYPE_ID = type_id, CREATED_DATE = created_date, DUE_DATE = due_date,
+      EXPIRY_DATE = expiry_date, PATH = path, DESCRIPTION = description
+      WHERE `SUPPORT_DOC`.SUPPORT_DOC_ID = support_doc_id;
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Update given support doc type
+-- -----------------------------------------------------
+-- procedure update_doc_type
+--    - update the given doc type, if it exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS update_doc_type;
 
 DELIMITER //
@@ -109,9 +190,21 @@ CREATE PROCEDURE `update_doc_type` (IN type_id INT
 , IN path VARCHAR(512)
 )
 BEGIN
-    UPDATE `TYPE`
-    SET TYPE_NAME = type_name, PATH = path
-    WHERE `TYPE`.type_id = type_id;
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(TYPE_ID) INTO checker
+    FROM `TYPE`
+    WHERE `TYPE`.TYPE_ID = type_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Document type does not exist.';
+    ELSE
+      UPDATE `TYPE`
+      SET TYPE_NAME = type_name, PATH = path
+      WHERE `TYPE`.type_id = type_id;
+    END IF;
 END //
 
 DELIMITER ;
